+++ Class17_2021_07_23/GeometricShapes/Picture.cs
@@ -32,7 +32,7 @@ namespace GeometricShapes
             }
         }
 
-        public void Resize(int mult)
+        public void Resize(double mult)
         {
             for (int i = 0; i < _amountOfFigures; i++)
             {
