+++ Class17_2021_07_23/GeometricShapes/Shapes/Line.cs
@@ -80,7 +80,7 @@ namespace GeometricShapes
             }
         }
 
-        public override void Resize(int mult)
+        public override void Resize(double mult)
         {
         }
