+++ Class17_2021_07_23/GeometricShapes/Shapes/PolyLine.cs
@@ -52,7 +52,7 @@ namespace GeometricShapes
             return points;
         }
 
-        public override void Resize(int mult)
+        public override void Resize(double mult)
         {
         }
